Complicate the artificial work further
It seemed possible that the compiler was managing to optimize aspects of what
was being done in conductArtificialWork such that it was completing within the
timestep.  With that in mind, I wondered if there was a way to complicate the
function further such that it would make optimization less likely

These additions try to make it less predictable how the result would have
changed.  They at least make it harder for me to predict it, but I can still see
a compiler being able to (given that I am still operating with what in Chapel
would be considered param values).

Still, the results fit without reaching "inf" in all the steps printed.  Which
was one of my goals.

----
Signed-off-by: Lydia Duncan <lydia-duncan@users.noreply.github.com>